3 - [Tech] Modules Management - move domain entities to its dir & add moshi
