+++ app/src/main/java/io/minoro75/demoweatherapp/remote/RemoteDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package io.minoro75.demoweatherapp.remote
 
-import io.minoro75.demoweatherapp.domain.forecast_entity.ForecastLocation
+import io.minoro75.demoweatherapp.forecast_entity.ForecastLocation
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
