+++ app/src/main/java/io/minoro75/demoweatherapp/remote/RemoteService.kt
@@ -1,8 +1,8 @@
 package io.minoro75.demoweatherapp.remote
 
 import io.minoro75.demoweatherapp.BuildConfig
-import io.minoro75.demoweatherapp.domain.city_name_entity.CityNameRoot
-import io.minoro75.demoweatherapp.domain.forecast_entity.ForecastRoot
+import io.minoro75.demoweatherapp.city_name_entity.CityNameRoot
+import io.minoro75.demoweatherapp.forecast_entity.ForecastRoot
 import retrofit2.http.GET
 import retrofit2.http.Query
