+++ app/src/main/java/io/minoro75/demoweatherapp/ui/city_weather/CityWeatherViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.minoro75.demoweatherapp.domain.forecast_entity.Weather
+import io.minoro75.demoweatherapp.forecast_entity.Weather
 import io.minoro75.demoweatherapp.repository.WeatherRepository
 import io.minoro75.demoweatherapp.utils.Resource
 import kotlinx.coroutines.flow.collect
