+++ app/src/main/java/io/minoro75/demoweatherapp/ui/city_weather/WeatherAdapter.kt
@@ -7,7 +7,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import io.minoro75.demoweatherapp.R
-import io.minoro75.demoweatherapp.domain.forecast_entity.Weather
+import io.minoro75.demoweatherapp.forecast_entity.Weather
 import java.text.SimpleDateFormat
 import java.util.Locale
