+++ domain/src/main/java/io/minoro75/demoweatherapp/forecast_entity/ForecastLocation.kt
@@ -1,4 +1,4 @@
-package io.minoro75.demoweatherapp.domain.forecast_entity
+package io.minoro75.demoweatherapp.forecast_entity
 
 import com.squareup.moshi.JsonClass
